All in on f64

+++ add.odin
@@ -27,7 +27,7 @@ add_mat_alloc :: proc(a, b: Matrix, allocator := context.allocator) -> (res: Mat
 	return res, nil
 }
 
-add_scalar_alloc :: proc(a: Matrix, s: f32, allocator := context.allocator) -> (res: Matrix, err: Matrix_Error) {
+add_scalar_alloc :: proc(a: Matrix, s: f64, allocator := context.allocator) -> (res: Matrix, err: Matrix_Error) {
 	res = alloc(a.rows, a.cols, allocator) or_return
 
 	for i in 0 ..< len(a.data) {
